fix: try to make suspense better

+++ packages/maskbook/src/plugins/ITO/hooks/usePoolPayload.ts
@@ -1,26 +1,26 @@
+import { useState } from 'react'
 import { PluginITO_RPC } from '../messages'
 import type { JSON_PayloadInMask } from '../types'
 
-const storage = new Map<string, JSON_PayloadInMask | Error>()
+const storage = new Map<string, JSON_PayloadInMask>()
 
 export function usePoolPayload(pid: string) {
-    if (!storage.has(pid)) throw suspender(pid)
-    return {
-        payload: storage.get(pid)!,
-        retry: () => retry(pid),
-    }
-}
+    const [e, set] = useState<Error | null>(null)
 
-async function retry(pid: string) {
-    storage.delete(pid)
-    if (!storage.has(pid)) throw suspender(pid)
-    return
+    const payload = storage.get(pid)
+    if (payload) {
+        return {
+            payload,
+            retry: () => {
+                storage.delete(pid)
+                throw suspender(pid).catch(set)
+            },
+        }
+    }
+    if (e) throw e
+    throw suspender(pid).catch(set)
 }
 
 async function suspender(pid: string) {
-    try {
-        storage.set(pid, await PluginITO_RPC.getPool(pid))
-    } catch (error) {
-        storage.set(pid, new Error())
-    }
+    storage.set(pid, await PluginITO_RPC.getPool(pid))
 }
